--- a/beacon_chain/gossip_processing/eth2_processor.nim
+++ b/beacon_chain/gossip_processing/eth2_processor.nim
@@ -250,6 +250,7 @@ proc aggregateValidator*(
   debug "Aggregate received", delay
 
   let v = self.attestationPool[].validateAggregate(
+      self.batchCrypto,
       signedAggregateAndProof, wallTime)
   if v.isErr:
     debug "Dropping aggregate",