… beacon_chain/gossip_processing/gossip_to_consensus.nim
… beacon_chain/gossip_processing/gossip_validation.nim
… beacon_chain/nimbus_beacon_node.nim
… beacon_chain/spec/crypto.nim
… beacon_chain/spec/signatures_batch.nim
… beacon_chain/validators/validator_duties.nim
… tests/test_attestation_pool.nim