# Batch attestations

This introduces attestations and aggregate batching for gossip validation.

At a low-level:
- an unaggregated attestation requires one cryptographic check
- an aggregated attestation requires 3 checks: selection proof, aggregator signature, aggregate signatures

The batching rules are:
- on a first attestation or aggregate, a 10ms timer is started
- A batch is processed after any of the following condition is reached:
  - the timer expired
  - Chronos is idle as defined by idleAsync so no other event (?)
  - 16 or more crypto checked are batched (6 aggregates can lead to 18 crypto check batched).

A single crypto check is at about 1ms on my machine (miller loop + final exponentiation).
A batch check is at 0.5ms per check
Constant overhead:
- Secure RNG: ~1ms
- Final exponentiation: ~0.5 ms

If a batch verification fails, checks are done individually.

Worst case expected latency would be with 18 checks, and 1 failing:
- 10ms idle timer
- 10.5ms for 18 checks
- 18ms for redoing the checks 1 by 1 if one failed

So about **40ms**.
Assuming a Rpi 4 is a conservative 4x less powerful than my CPU for cryptography, the conservative worst case latency is **160ms**.
In exchange **throughput is 2x higher**.

Given that starting from 9 checks (3 aggregates) the cost of batching is already amortized, it might be sensible to set the batch threshold to 8~12. We can also augment the idle timer to 20ms:
- 20 ms idle
- 5.5 ms for 9 checks
- 9 ms for redoing the checks 1 by 1 if one failed

**34.5ms** worst case latency, but with a longer timer so better batching opportunities.

However batching is interesting to handle bursts and 10ms is already enough in burst mode.