@mratsim 2021-03-21
## Bugfix

https://github.com/status-im/nimbus-eth2/pull/2439/commits/6dbb80c4a231ef91745bd29b4fe4cf509acb4387 introduces async validators.
The initial version of this PR had a tricky bug like https://github.com/status-im/nimbus-eth2/issues/1518 had a tricky bug with the stacktrace pointing to either chronos streams or posix read.

I suspect this was due in both cases to an async "reentrancy" bug. The P2P validators are sync proc wrapped into `async` here

https://github.com/status-im/nimbus-eth2/blob/baa43ae7f7f58807f2bf5308f8ceb420aae9d3fe/beacon_chain/networking/eth2_network.nim#L1591-L1617

Then in both cases async was needed within those P2P validators which caused issues:
- In the `sync` case to fetch ancestors (request_manager) and/or missing blocks (sync_manager)
- In this PR case to introduce deferred crypto verification to allow batching

## Performance

Aggregates require 3 crypto checks (proof, aggregator signature, aggregated signature)

On my machine it seems like we can do 18 crypto checks in just 10 ms
![image](https://user-images.githubusercontent.com/22738317/111918488-93ce8200-8a85-11eb-94d1-1c5482b22626.png)

Just 3 signatures or 9 signatures is about 1ms/sig
![image](https://user-images.githubusercontent.com/22738317/111918536-cd9f8880-8a85-11eb-857d-cfad6bc7b88a.png)