@cheatfate 2021-03-21
> The P2P validators are sync proc wrapped into async here
This is incorrect statement, if procedure do not perform `await` call it doesn't mean that its not async. All this validator callbacks are awaited by https://github.com/status-im/nim-libp2p/blob/70deac9e0d16f7d00a0d4404ed191042dbf079be/libp2p/protocols/pubsub/pubsub.nim#L483-L495. So validators are async procedures which are awaited and which returns immediate results.