@mratsim 2021-03-22
Some more benchmarks:

- 1 crypto check, this varies from 1ms to 3ms. I suspect the variation is when the RNG needs to refill entropy. Unsure if it's worth it to special case a path for it.
  
  ![image](https://user-images.githubusercontent.com/22738317/111959225-78e42800-8aee-11eb-89cb-b27105b7a158.png)

- 3 crypto checks. 2ms to 6ms (aggregates come with 3 checks)
  
  ![image](https://user-images.githubusercontent.com/22738317/111959429-bf398700-8aee-11eb-9ac8-2313ec76e7fa.png)
  ---------
  ![image](https://user-images.githubusercontent.com/22738317/111959678-10e21180-8aef-11eb-810c-ddde7a4d832e.png)

- 6 crypto checks (2 aggregates), consistent 2x speedup 2x3 checks (3.5 ms only)
- 9 crypto checks (3 aggregates), 5 ms only
- larger batch are about 0.5ms per additional crypto checks
  ![image](https://user-images.githubusercontent.com/22738317/111960304-dcbb2080-8aef-11eb-8ad8-ebadc85589c6.png)
  ![image](https://user-images.githubusercontent.com/22738317/111960511-26a40680-8af0-11eb-970b-7283768640ef.png)

- but there can be outliers
  ![image](https://user-images.githubusercontent.com/22738317/111960430-0a07ce80-8af0-11eb-9809-5470a3478ad3.png)