@cheatfate 2021-03-25
I think name of the PR confusing, because current PR apply batch processing only for gossipsub incoming attestations. It doesn't batch attestations in blocks (from gossip), and it doesnt batch attestations in blocks (from sync managers), and it doesn't batch attestations from (from json-rpc).