@mratsim 2021-03-26
> It doesn't batch attestations in blocks (from gossip), and it doesnt batch attestations in blocks (from sync managers)

Those are already batched https://github.com/status-im/nimbus-eth2/blob/8391d3f83e43e7e6d13b6bd716f06510e6ffd05c/beacon_chain/spec/signatures_batch.nim#L159-L174

Attestations from RPC can also use the same BatchCrypto. Actually even block batching can use it and it can work as a service but it's better to start small and focused before adding more risks and refactoring.